CMS-12052 Skip preflight check for CSR build
The CSR build does not like the fact that the react-sdk parent contains
a webpack dependency that is of a different version than it requries.
As the node_modules folder of the CSR contains the correct webpack
version, this is merely a warning that can be ignored.

We should enable the "preflight check" again once we built the SDK with
rollup.