CMS-12149 Include specific files in JAR instead of "everything except..."
On Jenkins, the local '.m2' folder is not excluded, so the local Maven
repository of the build (~19 MB) is included in every generic SNAPSHOT
JAR file.

Ensure that node_modules and build folders of the examples are removed
on "mvn clean" to avoid that local builds of the examples will end up
in the generated JAR file.

Also renamed 'eclude' in the Apache Rat config to 'exclude' (although
'eclude' actually seems to work! %-)